Added Thread::CurrentThreadId() method to wrap up thread id functionality in a more platform appropriate way.

+++ src/OpenThreads/win32/Win32ThreadPrivateData.h
@@ -54,7 +54,7 @@ private:
 
     HandleHolder tid;
 
-    int uniqueId;
+    size_t uniqueId;
 
     Affinity affinity;
